JS-22926 move auth logic to index.js
