+++ .gitignore
@@ -9,6 +9,7 @@ yarn-error.log
 # Build files
 lib
 module
+.idea/
 
 # Test files
 coverage
